use handler instead of read/write
